requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0

# RAG / Vector search
faiss-cpu>=1.7.4
openai>=1.0.0

# Profiling & utils
psutil>=5.9.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
